core,graph,server: Implement the update_query_prices RPC method
Update the query price for a specified subgraph.
Use u64 for query price, this is assuming that the price will be in
gwei denominated integers on the contracts.

+++ graph/src/data/subgraph/schema.rs
@@ -393,7 +393,7 @@ impl SubgraphDeploymentEntity {
 
     pub fn update_query_price_operations(
         id: &SubgraphDeploymentId,
-        price: BigDecimal,
+        price: u64,
     ) -> Vec<MetadataOperation> {
         let mut entity = Entity::new();
         entity.set("queryPrice", price);
